Merge pull request

+++ code/engine.vc2008/xrGame/level_sounds.cpp
@@ -213,7 +213,7 @@ void CLevelSoundManager::Update()
 			if (!indices.empty())
 			{
 				u32 idx			= Random.randI((u32)indices.size());
-				m_CurrentTrack	= indices[idx];
+				m_CurrentTrack	= (int)indices[idx];
 				SMusicTrack& T	= m_MusicTracks[m_CurrentTrack];
 				T.Play			();
 			}
